Extract a ForumPostScoped concern to share post loading and authorization
