+++ app/controllers/rooms/forums/posts/solutions_controller.rb
@@ -3,8 +3,11 @@
 # User#can_administer?, which allows the record's creator). The live UI update
 # rides on the model's tenant-scoped broadcasts, so the request just redirects.
 class Rooms::Forums::Posts::SolutionsController < ApplicationController
+  include ForumPostScoped
+
+  before_action :set_forum
   before_action :set_post
-  before_action :ensure_can_resolve
+  before_action :ensure_can_administer_post
 
   def create
     @post.solve!
@@ -15,15 +18,4 @@ class Rooms::Forums::Posts::SolutionsController < ApplicationController
     @post.reopen!
     redirect_back fallback_location: room_url(@forum)
   end
-
-  private
-    def set_post
-      @forum = Current.user.rooms.forums.find_by(id: params[:forum_id])
-      @post = @forum&.threads&.find_by(id: params[:post_id])
-      redirect_to root_url, alert: "Post not found or inaccessible" unless @post
-    end
-
-    def ensure_can_resolve
-      head :forbidden unless @post && Current.user.can_administer?(@post)
-    end
 end
