+++ app/controllers/rooms/forums/posts_controller.rb
@@ -2,9 +2,11 @@
 # reply-in-thread on an existing message): here the member authors a titled
 # opening post. Any forum member may post; membership is the gate.
 class Rooms::Forums::PostsController < ApplicationController
+  include ForumPostScoped
+
   before_action :set_forum
   before_action :set_post, only: %i[ edit update ]
-  before_action :ensure_can_edit, only: %i[ edit update ]
+  before_action :ensure_can_administer_post, only: %i[ edit update ]
 
   def create
     post = @forum.post!(title: post_params[:title], body: post_params[:body])
@@ -34,20 +36,6 @@ class Rooms::Forums::PostsController < ApplicationController
   end
 
   private
-    def set_forum
-      @forum = Current.user.rooms.forums.find_by(id: params[:forum_id])
-      redirect_to root_url, alert: "Forum not found or inaccessible" unless @forum
-    end
-
-    def set_post
-      @post = @forum&.threads&.find_by(id: params[:id])
-      redirect_to room_url(@forum), alert: "Post not found" unless @post
-    end
-
-    def ensure_can_edit
-      head :forbidden unless @post && Current.user.can_administer?(@post)
-    end
-
     def post_params
       params.require(:post).permit(:title, :body)
     end
